redraw image data into controlled bitmap context
an explicit redraw of the captured image into a bitmap context is required as QCPlugIn only supports ARGB8, BGRA8 and RGBAf while the source image may be in a different format - the 5D seems to be in RGBA8.

the redraw occurs in -execute:atTime:withArguments: which is not ideal, it should likely be moved to run after _didDownloadFile: or _didReadFile:, not block execute.

data is now acquired by reading the data into memory via _didReaFile: rather than downloading the image, then reading off disk through _didDownloadFile:. at some later point a quick comparison will be performe.

the bitmap context and its backing are released in callback when the provider is released, in a quick look, it does not appear as though allocations and releases are evenly paired, this may need to be looked into more closely.