wrap open session request with 1 sec timeout until it pulls the next available
ImageCaptureCore has an apparent bug leaving an open session request unanswered if two attempt to open it at nearly the same time (the second wins). using a 1 sec timeout on response, the patch can then move on to open a session on a different camera device should one be available.