disallow camera device selection when it already has a delegate
once a camera device is in use, another patch instance cannot select it. this is true for auto selection and for selection from the patch settings. this might prove to be problematic if the composition is selected in the finder and opened in the editor.

a later change may allow the camera device to be taken via selection in the popup but not via auto-selection.